requirements.txt
torch>=2.0
transformers>=4.40
sentence-transformers>=2.0
accelerate>=0.20
numpy>=1.24
scipy>=1.10
scikit-learn>=1.3
matplotlib>=3.7
datasets>=2.14
rank_bm25>=0.2
pyyaml>=6.0
tqdm>=4.65
networkx>=3.1
